stream settings: Add default_color field to the Stream Model.
Previously, Zulip's stream colors were generated independently for
each user (with an ordered sequence of colors). This is nice for the
onboarding experience, because it lets us avoid near-duplicate
colors. But for larger organizations, one can often want to use a
set of default colors(E.g. orange/red shades for alerts streams,
purple for social streams).

To address this added a `default_color` field in the `Stream` model,
in the same format of `Subscription.color`.  This is an optional
field and it defaults to None. This field intends to store the
default color of a stream.