stream settings: Add default_color to Stream Create Page.
Added `default_color` input to create stream page. if a user has not
picked a personal color for a stream, the default color of the stream
will be used.

Added a new spectrum creation function to support stream default
colors named `set_colorpicker_default_color` in `stream_color.js`.
this new type of spectrum creation is due to different params we
pass to stream creation spectrum ie(allowEmpty,clickoutFiresChange)
and modified `show()`,`hide()`,`change()`.

If the value in the colorpicker field is null, then the default
color field wont be passed in the POST Data, else if the user has
choosen a color,The field will be passed.

I also added the default_color field to optional_keys in add_subs
schema in `views/streams.py`, and updated the stream creation
functions in `lib/streams.py` to accept this new field.